Separate CI and CD
As the step condition could rely only on the ref name, that could be a
tag or branch, there's no way to effectively publish only versions that
were integrated on master. By separating the workflows we can rely on
`on` branch and tag filters, it introduces a little of duplication but
that's no big deal.